Added "send_info.py" script
Dirty script to send basic information to a web server in the form of a GET request. Written to be later disguised as a text file called "if_found.txt" for social engineering exercise. Remote web server logs can be checked to see whether host information was sent to it.